Prevent listing huge buckets on S3.
Previously, in when `AWS_PRELOAD_METADATA` is True,
`S3BotoStorage.entries` would list every file in the bucket. However,
if there is a `location` specified on the storage, we don't use the
whole storage bucket with that storage backend. So pass our `location`
as the `prefix` to the call to `entries`.